requirements.txt
streamlit
langchain_community
langchain_core
python-dotenv